+++ src/bcra_sdk/models/__init__.py
@@ -1,4 +1,5 @@
 from .cheques import ResultGetChequesRechazadosV1
+from .cotizaciones import ResultGetCotizacionesV1
 from .denunciados import ResultGetChequeDenunciadoV1
 from .deudores import (
     EntidadHistorica,
@@ -14,6 +15,7 @@ __all__ = [
     "PeriodoHistorica",
     "ResultGetChequeDenunciadoV1",
     "ResultGetChequesRechazadosV1",
+    "ResultGetCotizacionesV1",
     "ResultGetDeudasHistoricasV1",
     "ResultGetDeudasV1",
     "ResultGetDivisasV1",